extract EditTicketFormModal into new NewTicketFormModal component
