+++ assets/src/editor/events/tickets/editor-ticket/edit-form/index.js
@@ -1,4 +1,8 @@
 export { default as EditTicketForm } from './edit-ticket-form';
 export { default as EditTicketFormModal } from './edit-ticket-form-modal';
+export { default as NewTicketFormModal } from './new-ticket-form-modal';
 export { default as useTicketEditorId } from './use-ticket-editor-id';
-export { default as useTicketEntityFormInputPrefix } from './use-ticket-editor-id';
+export { default as useTicketEntityFormInputPrefix }
+	from './use-ticket-editor-id';
+export { default as useTicketFormLoadingState }
+	from './use-ticket-form-loading-state';
